Create and connect to Room
