+++ src/App/index_to_rename.tsx
@@ -1,2 +1,3 @@
 import App from "./App";
+
 export default App;
