+++ src/pages/Home/Home.tsx
@@ -1,14 +1,36 @@
-import React from "react";
+import Box from "@material-ui/core/Box";
+import Button from "@material-ui/core/Button";
 import Container from "@material-ui/core/Container";
-import { Box, Typography, Button } from "@material-ui/core";
-import { Link } from "react-router-dom";
+import Typography from "@material-ui/core/Typography";
+import React from "react";
+import { useHistory } from "react-router-dom";
+
+import { useApi } from "../../services/api";
 
 import Styles from "./Home.module.scss";
 
 const Home: React.FC = () => {
+  const api = useApi();
+  const history = useHistory();
+
+  const onCreateRoomClick = async (): Promise<void> => {
+    try {
+      // Create Room
+      const res = await api.room.createRoom();
+
+      // Go to room
+      const roomUuid = res.data.room.uuid;
+      history.push(`/rooms/${roomUuid}`);
+    } catch (error) {
+      // TODO: Provide real feedback
+      console.error(error.response);
+    }
+  };
+
   return (
     <Container className={Styles.HomeContainer}>
       <h1>Impact Title</h1>
+
       <Box marginTop={10}>
         <Typography>
           Quae dum ita struuntur, indicatum est apud Tyrum indumentum regale
@@ -18,9 +40,10 @@ const Home: React.FC = () => {
           multi, qui atrocium criminum ponderibus urgebantur.
         </Typography>
       </Box>
-      <Link to="/room">
-        <Button color="primary">Create Room</Button>
-      </Link>
+
+      <Button color="primary" onClick={onCreateRoomClick}>
+        Create Room
+      </Button>
     </Container>
   );
 };
