+++ src/pages/Home/index.tsx
@@ -1,2 +1,3 @@
 import Home from "./Home";
+
 export default Home;
