+++ src/pages/Login/index.tsx
@@ -1,2 +1,3 @@
 import Login from "./Login";
+
 export default Login;
