+++ src/pages/Register/index.tsx
@@ -1,2 +1,3 @@
 import Register from "./Register";
+
 export default Register;
