--- src/pages/Room/AddUrlDialog/AddUrlDialog.module.scss
@@ -1,3 +1,0 @@
-.AddMusic {
-  margin: 20px;
-}
