+++ src/pages/RoomPage/AddUrlDialog/AddUrlDialog.module.scss
@@ -1,0 +1,3 @@
+.AddMusic {
+  margin: 20px;
+}
