+++ src/pages/RoomPage/RoomPage.module.scss
@@ -1,0 +1,15 @@
+.ytPlayer {
+  width: 100%;
+  height: 60vh;
+}
+
+.RoomContainer {
+  width: 100%;
+  margin-top: 30px;
+  height: 60vh;
+}
+
+.NextLabel {
+  display: flex;
+  margin: 20px;
+}
