+++ src/pages/RoomPage/index.tsx
@@ -1,0 +1,3 @@
+import RoomPage from "./RoomPage";
+
+export default RoomPage;
