--- src/types/MusicItem.ts
@@ -1,4 +1,0 @@
-export type MusicItem = {
-  id: string;
-  url: string;
-};
