+++ src/types/RoomMember.ts
@@ -1,0 +1,5 @@
+export default interface RoomMember {
+  id: string;
+  name: string;
+  isConnected: boolean;
+}
