+++ src/types/RoomPlaylist.ts
@@ -1,0 +1,8 @@
+import RoomPlaylistItem from "./RoomPlaylistItem";
+import RoomPlaylistCurrent from "./RoomPlaylistCurrent";
+
+export default interface RoomPlaylist {
+  current?: RoomPlaylistCurrent;
+  nextItems: RoomPlaylistItem[];
+  isPlaying: boolean;
+}
