+++ src/types/RoomPlaylistCurrent.ts
@@ -1,0 +1,7 @@
+import RoomPlaylistItem from "./RoomPlaylistItem";
+
+export default interface RoomPlaylistCurrent {
+  item: RoomPlaylistItem;
+  playingSince: Date;
+  musicOffset: number;
+}
