+++ src/types/RoomPlaylistItem.ts
@@ -1,0 +1,5 @@
+export default interface RoomPlaylistItem {
+  videoId: string;
+  uuid: string;
+  addedBy: string;
+}
