+++ src/utils/envOrThrow.ts
@@ -1,15 +1,15 @@
 export default function envOrThrow(
-    key: string,
-    defaultValue: string | undefined = undefined,
-  ): string {
-    const val = process.env[key];
+  key: string,
+  defaultValue: string | undefined = undefined
+): string {
+  const val = process.env[key];
 
-    if (!val) {
-      if (!defaultValue) {
-        throw new Error(`Missing environment variable "${key}"`);
-      }
-      return defaultValue;
+  if (!val) {
+    if (!defaultValue) {
+      throw new Error(`Missing environment variable "${key}"`);
     }
-
-    return val;
+    return defaultValue;
   }
+
+  return val;
+}
